==> code/R/IMData_analysis_3/10_visualize_predicted_clusters_per_metadata.Rmd <==
---
title: "Visualize predicted clusters per metadata group"
output: html_document
date: '2022-08-15'
---

```{r setup, include=FALSE, cache=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "120%")
```

## Load data

We will first read in the SPE for tumor and non-tumor cells with the predicted cell type based on scRNAseq reference datasets. As we weren't sure whether the schwann cell clusters are tumor cells or not, we predicted their cell type using both reference datasets (Jansky and Lee). Upon closer examination, we decided to assign them to the tumor cell metacluster, because a small proportion of the cluster might contain SCP or cycling SCPs. 

```{r read-data-pheno, message=FALSE}

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"

# For separate tumor and non-tumor prediction (e.g. Jansky-Lee)
spe_tumor <- readRDS(file.path(path,"results/public_datasets/Jansky/results/spe_predicted_tumor_clusters.rds"))
spe_tumor$predicted_clusters <- spe_tumor$predicted_tumor_clusters
spe_tumor$predicted_tumor_clusters <- NULL

spe_non_tumor <- readRDS(file.path(path,"results/public_datasets/Lee/results/spe_immune_clusters.rds"))
spe_non_tumor$predicted_clusters <- spe_non_tumor$predicted_immune_clusters
spe_non_tumor$predicted_immune_clusters <- NULL
#spe_non_tumor <- spe_non_tumor[,spe_non_tumor$celltype!="78_schwann cells" & spe_non_tumor$celltype !="78_SCP"]

spe <- cbind(spe_tumor, spe_non_tumor)
#spe[,spe$celltype=="78_schwann cells" | spe$celltype =="78_SCP"]$metacluster <- "tumor"

#For Fetahu
# spe <- readRDS(file.path(path,"public_datasets/Fetahu/results/spe_predicted_clusters.rds"))


saveRDS(spe, file.path(path, "results/public_datasets/Jansky_Lee/spe_predicted-clusters_Jansky-Lee_2.RDS"))
```

Next, we will plot the predicted celltypes in the UMAP space. 

```{r cell type umap, fig.width=30, fig.height=15, message=FALSE}
library(dittoSeq)
library(RColorBrewer)

set.seed(230714)
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]
col_vector_74 = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))
col_vector_433 = colors()[grep('gr(a|e)y', grDevices::colors(), invert = T)]

col_predicted_labels = setNames(sample(col_vector_433, length(unique(spe$predicted_clusters))), unique(spe$predicted_clusters))

## UMAP colored by cell type and expression - dittoDimPlot
p2 <- dittoDimPlot(spe, var = "predicted_clusters", 
             reduction.use = "UMAP", size = 1,
             do.label = TRUE,
             legend.show=T) +
  scale_color_manual(values = col_predicted_labels) +
  theme(legend.title = element_blank()) +
  ggtitle("Inferrred clusters on UMAP")

p2

```
Now, we will plot the clusters separately in the dim-red space per PT and BM. 

```{r visualizing-umap-by-marker, message=F, fig.width=20}

celltype <- spe$predicted_clusters
mat_red_df <- as.data.frame(reducedDim(spe, "UMAP"))
mat_red_df["tissue"] <- spe$tissue

samples <- unique(celltype)
col_tissue = c("PT" = "deepskyblue", "BM" = "orange")

for(i in 1: length(samples)){

  sample_binary <- celltype == samples[i] 
  
  background <- as.data.frame(mat_red_df[!sample_binary,])
  foreground <- as.data.frame(mat_red_df[sample_binary,])
  plot <- ggplot(background, aes(x=V1, y=V2) ) +
    geom_point(color="gray90", size=0.02) +
    geom_point(data=foreground, aes(x=V1, y=V2, colour=tissue), show.legend=F, size=0.02, alpha=0.25) +
    scale_colour_manual(values = col_tissue) +
    ggtitle(samples[i]) +
    theme(plot.title = element_text(size = 5), 
          axis.title = element_text(size=5), 
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(), 
          panel.background = element_blank(), 
          axis.line = element_line(colour = "black"))
  
  ggsave(filename=file.path(path, paste0("results/UMAP/Fetahu/", samples[i] , ".tiff")), width = 10, height = 5, units = "in", device='tiff', dpi=500)
}

```

Next, we will visualize the cluster per metadata group as described [here](https://r-graph-gallery.com/48-grouped-barplot-with-ggplot2.html).
```{r barplot-cluster-per-sample, fig.height=5, fig.width=4}
 # library
library(ggplot2)
library(tidyverse)

#change according to group of interest
# spe_reduced <- spe[,spe$tissue=="BM" & spe$control=="no" & spe$metacluster!="tumor"]
# spe_reduced <- spe_reduced[,spe_reduced$MYCN_amp==1 | spe_reduced$MYCN_amp==3]
# spe_reduced[,spe_reduced$MYCN_amp==3]$MYCN_amp <- 1
# groupname <- "timepoint"

spe_reduced <- spe[,spe$control=="no" & spe$metacluster=="tumor" & spe$tissue=="BM" & spe$timepoint=="DE"]
spe_reduced <- spe_reduced[,spe_reduced$MYCN_amp==0 | spe_reduced$MYCN_amp==1 | spe_reduced$MYCN_amp==3]
spe_reduced[,spe_reduced$MYCN_amp==3]$MYCN_amp <- 1
groupname <- "MYCN_amp"

# spe_reduced <- spe[,spe$control=="no" & spe$metacluster=="tumor"]
# spe_reduced <- spe_reduced[,spe_reduced$TERT_rear==0 | spe_reduced$TERT_rear==1]
# groupname <- "TERT_rear"

# spe_reduced <- spe[,spe$tissue=="BM"&spe$metacluster!="tumor"]
# groupname <- "control"

# spe_reduced <- spe[,spe$timepoint=="DE"  & spe$metacluster!="tumor"]
# spe_reduced <- spe_reduced[,spe_reduced$MYCN_amp==0]
# groupname <- "tissue"

group <- colData(spe_reduced)[,groupname]

#col_celltype <- col_celltype[tumor_celltypes]
#col_celltype <- col_celltype[order(as.integer(sub('_.*', '', names(col_celltype))), decreasing=F)]


celltype <- spe_reduced$predicted_clusters
tumor_celltypes <- c("late SCPs", "cycling SCPs", "SCPs", "Bridge", "connecting Chromaffin cells", "Chromaffin cells", "late Chromaffin cells", "cycling Neuroblasts", "Neuroblasts", "late Neuroblasts")


#col_celltype <- col_celltype[order(names(col_celltype))]
col_celltype <- col_predicted_labels[unique(spe_reduced$predicted_clusters)]
data <- data.frame(sample = group, celltype)
data$freq <- 1

data_wide <- data %>% pivot_wider(names_from = c("celltype"), values_from = "freq", values_fn = sum, values_fill = 0)

data_long <- data_wide %>% pivot_longer(cols = !sample, names_to = "celltype", values_to = "count") 
data_long$celltype <- factor(data_long$celltype, levels=tumor_celltypes)

#write.csv(data_wide, file.path(path, paste0("results/metadata/predicted_clusters/nMNA_celltype-per-", groupname, "_non-tumor.csv")))
 
# Stacked + percent
ggplot(data_long, aes(fill=celltype, y=count, x=sample)) + 
  ggtitle(paste0("BM_celltype per ", groupname)) +
  geom_bar(position="fill", 
    #position="stack",
    stat="identity",
    show.legend = T) + 
  scale_fill_manual(values = col_celltype) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) 
```

## Session Info

<details>
   <summary>SessionInfo</summary>
   
```{r, echo = FALSE}

sessionInfo()
```
</details>
