==> code/R/IMData_analysis_3/12_pseudotime_scorpion_in-house.Rmd <==
---
---
title: "Investigating maturity of identified tumor cells"
author: "Robrecht Cannoodt"
---

<!-- github markdown built using 
rmarkdown::render("vignettes/ginhoux.Rmd", output_format = "github_document")
-->

```{r, setseed, echo=FALSE}
set.seed(1)
```

In this vignette, SCORPIUS is used to infer a trajectory through tumor cells.

```{r load_data, message=FALSE}
library(SCORPIUS)
library(SpatialExperiment)

path <- "/mnt/Multimodal_Imaging_Daria/PIPELINE/NEWEST_RESULTS"
spe <- readRDS(file.path(path,"results/in_house/spe_clustered_5.rds"))
spe <- spe[,spe$metacluster=="tumor"]
expression <- t(assay(spe, "exprs")[rowData(spe)$use_channel,])
```

We will look at the first six cells and the expression of the first 6 markers.
```{r show_expression}
expression[1:6, 1:6]
```

`sample_info` is a data frame with the metadata of the cells, containing cell types of the individual cells.
```{r show_sample_info}
sample_info <- data.frame(celltype=spe$celltype,row.names=colnames(spe))
head(sample_info)

```

In order to infer a trajectory through this data, SCORPIUS first reduces the dimensionality of the dataset.


## Reduce dimensionality of the dataset
SCORPIUS uses Torgerson multi-dimensional scaling to reduce the dataset to three dimensions.
This technique attempts to place the cells in a space such that the distance between any
two points in that space approximates the original distance between the two cells as well as possible.

The distance between any two samples is defined as their correlation distance, namely `1 - (cor(x, y)+1)/2`.
The reduced space is constructed as follows:
```{r perform_mds}

space <- reduce_dimensionality(expression, "spearman", ndim = 3)
```

The new space is a `r nrow(space)`-by-`r ncol(space)` matrix, and can be visualised with or without colouring of the
different cell types.
```{r show_dimred}
draw_trajectory_plot(space, progression_group = as.factor(sample_info$celltype), contour = TRUE)
```


## Inferring a trajectory through the cells
The main goal of SCORPIUS is to infer a trajectory through the cells, and orden the cells according to the inferred timeline.

SCORPIUS infers a trajectory through several intermediate steps, which are all executed as follows:
```{r infer_trajectory}
traj <- infer_trajectory(space)
saveRDS(traj, file.path(path, "results/in_house/pseudotime/inferred_traj.rds"))
```


The result is a list containing the final trajectory `path` and the inferred timeline for each sample `time`.

The trajectory can be visualised with respect to the samples by passing it to `draw_trajectory_plot`:
```{r plot_trajectory}

draw_trajectory_plot(
  space, 
  progression_group = as.factor(sample_info$celltype),
  path = traj$path,
  contour = TRUE,
  point_size=0.5,
  point_alpha=1,
  progression_group_palette=metadata(spe)$color_vectors$col_celltype[unique(spe[,spe$metacluster=="tumor"]$celltype)]
)
```


## Finding candidate marker genes
We search for genes whose expression is seems to be a function of the trajectory timeline that was inferred, 
as such genes might be good candidate marker genes for dendritic cell maturation.

```{r find_tafs}
gimp <- gene_importances(expression, traj$time, num_permutations = 0, num_threads = 8)
gene_sel <- gimp
expr_sel <- expression[,gene_sel$gene]
```

```{r echo=F}
# reverse the trajectory. This does not change the results in any way,
# other than the heatmap being ordered more logically.
# traj <- reverse_trajectory(traj)
```

To visualise the expression of the selected genes, use the `draw_trajectory_heatmap` function.
```{r visualise_tafs, fig.keep='first', fig.width=20, fig.height=10}
rand_cells <- sample(seq_len(nrow(expr_sel)), 1000)
draw_trajectory_heatmap(expr_sel[rand_cells,], traj$time[rand_cells], as.factor(sample_info$celltype)[rand_cells], show_labels_row=T, progression_group_palette=metadata(spe)$color_vectors$col_celltype[unique(spe[,spe$metacluster=="tumor"]$celltype)], scale_features=F)
```

```{r moduled_tafs, fig.keep='first', fig.width=20, fig.height=15}

modules <- extract_modules(scale_quantile(expr_sel), traj$time, verbose = FALSE)
draw_trajectory_heatmap(expr_sel, traj$time, as.factor(sample_info$celltype), show_labels_row=T, progression_group_palette=metadata(spe)$color_vectors$col_celltype[unique(spe[,spe$metacluster=="tumor"]$celltype)], modules=modules)
```
